new file:   CMakeLists.txt     new file:   build/ALL_BUILD.vcxproj 	new file:   build/ALL_BUILD.vcxproj.filters 	new file:   build/CG1-Engine.sln 	new file:   build/CMakeCache.txt 	new file:   build/CMakeFiles/3.29.8/CMakeCCompiler.cmake 	new file:   build/CMakeFiles/3.29.8/CMakeCXXCompiler.cmake 	new file:   build/CMakeFiles/3.29.8/CMakeDetermineCompilerABI_C.bin 	new file:   build/CMakeFiles/3.29.8/CMakeDetermineCompilerABI_CXX.bin 	new file:   build/CMakeFiles/3.29.8/CMakeRCCompiler.cmake 	new file:   build/CMakeFiles/3.29.8/CMakeSystem.cmake 	new file:   build/CMakeFiles/3.29.8/CompilerIdC/CMakeCCompilerId.c 	new file:   build/CMakeFiles/3.29.8/CompilerIdC/CompilerIdC.exe 	new file:   build/CMakeFiles/3.29.8/CompilerIdC/CompilerIdC.vcxproj 	new file:   build/CMakeFiles/3.29.8/CompilerIdC/Debug/CMakeCCompilerId.obj 	new file:   build/CMakeFiles/3.29.8/CompilerIdC/Debug/CompilerIdC.exe.recipe 	new file:   build/CMakeFiles/3.29.8/CompilerIdC/Debug/CompilerIdC.tlog/CL.command.1.tlog 	new file:   build/CMakeFiles/3.29.8/CompilerIdC/Debug/CompilerIdC.tlog/CL.read.1.tlog 	new file:   build/CMakeFiles/3.29.8/CompilerIdC/Debug/CompilerIdC.tlog/CL.write.1.tlog 	new file:   build/CMakeFiles/3.29.8/CompilerIdC/Debug/CompilerIdC.tlog/Cl.items.tlog 	new file:   build/CMakeFiles/3.29.8/CompilerIdC/Debug/CompilerIdC.tlog/CompilerIdC.lastbuildstate 	new file:   build/CMakeFiles/3.29.8/CompilerIdC/Debug/CompilerIdC.tlog/link.command.1.tlog 	new file:   build/CMakeFiles/3.29.8/CompilerIdC/Debug/CompilerIdC.tlog/link.read.1.tlog 	new file:   build/CMakeFiles/3.29.8/CompilerIdC/Debug/CompilerIdC.tlog/link.secondary.1.tlog 	new file:   build/CMakeFiles/3.29.8/CompilerIdC/Debug/CompilerIdC.tlog/link.write.1.tlog 	new file:   build/CMakeFiles/3.29.8/CompilerIdCXX/CMakeCXXCompilerId.cpp 	new file:   build/CMakeFiles/3.29.8/CompilerIdCXX/CompilerIdCXX.exe 	new file:   build/CMakeFiles/3.29.8/CompilerIdCXX/CompilerIdCXX.vcxproj 	new file:   build/CMakeFiles/3.29.8/CompilerIdCXX/Debug/CMakeCXXCompilerId.obj 	new file:   build/CMakeFiles/3.29.8/CompilerIdCXX/Debug/CompilerIdCXX.exe.recipe 	new file:   build/CMakeFiles/3.29.8/CompilerIdCXX/Debug/CompilerIdCXX.tlog/CL.command.1.tlog 	new file:   build/CMakeFiles/3.29.8/CompilerIdCXX/Debug/CompilerIdCXX.tlog/CL.read.1.tlog 	new file:   build/CMakeFiles/3.29.8/CompilerIdCXX/Debug/CompilerIdCXX.tlog/CL.write.1.tlog 	new file:   build/CMakeFiles/3.29.8/CompilerIdCXX/Debug/CompilerIdCXX.tlog/Cl.items.tlog 	new file:   build/CMakeFiles/3.29.8/CompilerIdCXX/Debug/CompilerIdCXX.tlog/CompilerIdCXX.lastbuildstate 	new file:   build/CMakeFiles/3.29.8/CompilerIdCXX/Debug/CompilerIdCXX.tlog/link.command.1.tlog 	new file:   build/CMakeFiles/3.29.8/CompilerIdCXX/Debug/CompilerIdCXX.tlog/link.read.1.tlog 	new file:   build/CMakeFiles/3.29.8/CompilerIdCXX/Debug/CompilerIdCXX.tlog/link.secondary.1.tlog 	new file:   build/CMakeFiles/3.29.8/CompilerIdCXX/Debug/CompilerIdCXX.tlog/link.write.1.tlog 	new file:   build/CMakeFiles/3.29.8/VCTargetsPath.txt 	new file:   build/CMakeFiles/3.29.8/VCTargetsPath.vcxproj 	new file:   build/CMakeFiles/3.29.8/VCTargetsPath/x64/Debug/VCTargetsPath.recipe 	new file:   build/CMakeFiles/3.29.8/VCTargetsPath/x64/Debug/VCTargetsPath.tlog/VCTargetsPath.lastbuildstate     new file:   build/CMakeFiles/CMakeConfigureLog.yaml 	new file:   build/CMakeFiles/TargetDirectories.txt 	new file:   build/CMakeFiles/cmake.check_cache 	new file:   build/CMakeFiles/e38153d04174ba18ea71700bbeae1985/generate.stamp.rule 	new file:   build/CMakeFiles/generate.stamp 	new file:   build/CMakeFiles/generate.stamp.depend 	new file:   build/CMakeFiles/generate.stamp.list 	new file:   build/Debug/cg1.exe 	new file:   build/Debug/cg1.pdb     new file:   build/ZERO_CHECK.vcxproj 	new file:   build/ZERO_CHECK.vcxproj.filters 	new file:   build/cg1.dir/Debug/cg1.exe.recipe 	new file:   build/cg1.dir/Debug/cg1.ilk 	new file:   build/cg1.dir/Debug/cg1.tlog/CL.command.1.tlog 	new file:   build/cg1.dir/Debug/cg1.tlog/CL.read.1.tlog 	new file:   build/cg1.dir/Debug/cg1.tlog/CL.write.1.tlog 	new file:   build/cg1.dir/Debug/cg1.tlog/Cl.items.tlog 	new file:   build/cg1.dir/Debug/cg1.tlog/CustomBuild.command.1.tlog 	new file:   build/cg1.dir/Debug/cg1.tlog/CustomBuild.read.1.tlog 	new file:   build/cg1.dir/Debug/cg1.tlog/CustomBuild.write.1.tlog 	new file:   build/cg1.dir/Debug/cg1.tlog/cg1.lastbuildstate 	new file:   build/cg1.dir/Debug/cg1.tlog/link.command.1.tlog 	new file:   build/cg1.dir/Debug/cg1.tlog/link.read.1.tlog 	new file:   build/cg1.dir/Debug/cg1.tlog/link.secondary.1.tlog 	new file:   build/cg1.dir/Debug/cg1.tlog/link.write.1.tlog 	new file:   build/cg1.dir/Debug/main.obj 	new file:   build/cg1.dir/Debug/vc143.pdb 	new file:   build/cg1.vcxproj 	new file:   build/cg1.vcxproj.filters 	new file:   build/cmake_install.cmake 	new file:   build/x64/Debug/ALL_BUILD/ALL_BUILD.recipe 	new file:   build/x64/Debug/ALL_BUILD/ALL_BUILD.tlog/ALL_BUILD.lastbuildstate 	new file:   build/x64/Debug/ALL_BUILD/ALL_BUILD.tlog/CustomBuild.command.1.tlog 	new file:   build/x64/Debug/ALL_BUILD/ALL_BUILD.tlog/CustomBuild.read.1.tlog# 	new file:   build/x64/Debug/ALL_BUILD/ALL_BUILD.tlog/CustomBuild.write.1.tlog 	new file:   build/x64/Debug/ZERO_CHECK/ZERO_CHECK.recipe 	new file:   build/x64/Debug/ZERO_CHECK/ZERO_CHECK.tlog/CustomBuild.command.1.tlog 	new file:   build/x64/Debug/ZERO_CHECK/ZERO_CHECK.tlog/CustomBuild.read.1.tlog 	new file:   build/x64/Debug/ZERO_CHECK/ZERO_CHECK.tlog/CustomBuild.write.1.tlog 	new file:   build/x64/Debug/ZERO_CHECK/ZERO_CHECK.tlog/ZERO_CHECK.lastbuildstate 	modified:   engine/camera.h 	modified:   engine/light.h 	modified:   engine/ray.h 	modified:   engine/shapes/material.h 	modified:   engine/shapes/plane.h 	modified:   engine/shapes/shape.h 	modified:   engine/shapes/sphere.h 	modified:   main.cpp 	modified:   utils/vec3.h

+++ engine/light.h
@@ -7,9 +7,9 @@ class Light {
     public:
         Vec3 pos;
         Vec3 color;
-        float intensity;
+        double intensity;
         Light (): pos(Vec3()), color(Vec3(1.0, 1.0, 1.0)), intensity(1.0) {}
-        Light (Vec3 pos, Vec3 color, float intensity): pos(pos), color(color), intensity(intensity) {}
+        Light (Vec3 pos, Vec3 color, double intensity): pos(pos), color(color), intensity(intensity) {}
 };
 
 #endif
